Integrate Unreal vehicle sensor demo into main example
